requirements.txt
# Core FastAPI and ASGI server
fastapi==0.128.0
uvicorn[standard]==0.40.0
python-multipart==0.0.21

# PaddleOCR and PaddlePaddle
paddleocr==3.3.2
paddlepaddle==3.3.0

# Image processing
pillow==12.1.0
opencv-python-headless==4.12.0.88

# HTTP client for URL downloads
aiohttp==3.13.3
httpx==0.28.1

# Configuration and validation
pydantic==2.12.5
pydantic-settings==2.12.0
python-dotenv==1.2.1

# Utilities
numpy==2.4.1
